Function Index

JuMPTutorials.jl

JuMPTutorials.weave_all

JuMPTutorials.weave_all_f

JuMPTutorials.weave_file

JuMPTutorials.weave_file_f

JuMPTutorials.weave_folder

JuMPTutorials.weave_folder_f

JuMPTutorials.weave_all — Method.

weave_all()

Checks every tutorial for modifications and updates the notebook and tests accordingly.

source

JuMPTutorials.weave_all_f — Method.

weave_all_f()

Use Weave to convert every tutorial irrespective of whether it has been modified or not.

source

JuMPTutorials.weave_file — Method.

weave_file(folder::AbstractString, file::AbstractString)

Checks if the file has been modified since the last Weave and updates the notebook and tests accordingly.

folder = Name of the folder the tutorial is in

file = Name of the tutorial file

source

JuMPTutorials.weave_file_f — Method.

weave_file_f(folder::AbstractString, file::AbstractString)

Use Weave to convert the given file irrespective of whether it has been modified or not.

folder = Name of the folder the tutorial is in

file = Name of the tutorial file

source

JuMPTutorials.weave_folder — Method.

weave_file(folder::AbstractString)

Checks the files present in the specified folder for modifications and updates the notebook and tests accordingly.

folder = Name of the folder to check

source

JuMPTutorials.weave_folder_f — Method.

weave_folder_f(folder::AbstractString, file::AbstractString)

Use Weave to convert every file in the specified folder irrespective of whether it has been modified or not.

folder = Name of the folder the tutorial is in

source